pyproject.toml
[tool.poetry]
name = "ai-financial-agent"
version = "0.1.0"
description = "An AI-powered financial agent team using LangGraph and Claude"
authors = ["You"]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.4"
langchain-anthropic = "^0.3"
langchain-core = "^0.3"
pydantic = "^2.0"
rich = "^13.0"
python-dotenv = "^1.0"
requests = "^2.31"
yfinance = "^1.2"
fastapi = "^0.115"
uvicorn = {version = "^0.34", extras = ["standard"]}
jinja2 = "^3.1"
sse-starlette = "^2.0"
python-multipart = "^0.0.20"
pymupdf = "^1.24"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
